Fix to infinite loop condition when writing calendar
There was a problem where a call to getEndTime() would call
m_calendar.setTime(), when this call happened in the while loop as well.
Because the loop terminal condition called m_calendar.getTimeInMillis(),
if the last range.getEnd() is some date more than a day before
mpxjException.getToDate(), then the loop would never terminate. This is
fixed by creating a new calendar instance in the scope of the loop so
that m_calendar isn't being used in the terminal condition when it can
accidentally be set in a getter method.