Symmetry between Primavera PM reader/writer
Introduced method reversePercentage() in the reader, the functional
inverse of getPercentage() in the writer, including null handling, which
wasn't in the reader before; calling
row.getPercentComplete().doubleValue() would result in a
NullPointerException if the xml didn't contain the field number. In
addition, resource costs added in the writer to match the section in the
reader which looks for these fields.