UDF support for Primavera PM format
Added custom field support using existing custom field classes, as well
as some modifications. One big modification is the handling of UDFs in
the reader; because there is no deterministic way to map the object IDs
of the user fields that Primavera exports to FieldTypes in field
containers, the UDF data is simply left as raw data in the
CustomFieldContainer accessible from the ProjectFile returned by the
reader.