Alter method used to get temp filename
The Tempfile.new method generates a unique temporary filename and
creates the file.  On Windows, Ruby apparently does not maintain a lock
on the file because the call to java on line 22 is able to write to the
file.  It does maintain the lock under Linux, BSD, and MacOS and java
is not able to write to the file.  Tempfile.new uses
Dir::Tmpname.make_tmpname to generate the filename, so just use it.