Preliminary fix for PMFileWriter.
Until now the PrimaveraPMFileWriter only uses the Tasks getSummary()-Method to check if should be
written as Activity or WBS into the xml file.
However, this does not work with empty WBS.
This preliminary fix uses the task_type to distinguish between WBS (task_type is null) and
Activities (task_type is set). However, this does not work if a xml file has been used for import,
because then the task_type is not read.